Move sub-classes of "FlowLayout" class to their own files
